Convert Doppler Bins to Normalized Doppler Frequency

This function retrieves the normalized Doppler frequencies corresponding to the specified bins in a given SeaSondeR object.

Usage

seasonder_Bins2NormalizedDopplerFreq(seasonder_cs_object, bins)

Arguments

seasonder_cs_object

A SeaSondeR cross-spectral object containing Doppler bin metadata.

bins

A numeric vector specifying the Doppler bin indices.

Value

A numeric vector of normalized Doppler frequencies corresponding to the specified bins.

Details

This function first retrieves the Doppler bin frequencies in normalized form using seasonder_getDopplerBinsFrequency. It then selects the normalized Doppler frequencies corresponding to the specified bin indices.

Normalized Doppler Frequency Calculation: The normalized Doppler frequency is typically defined as: $$f_{norm} = \frac{f_{doppler}}{f_{bragg}}$$ where:

\(f_{norm}\) is the normalized Doppler frequency,

\(f_{doppler}\) is the Doppler frequency of a given bin,

\(f_{bragg}\) is the Bragg frequency, computed based on radar wavelength.

See also

Examples

# Read a sample Cross-Spectral (CS) file to create a SeaSondeRCS object
specs_path <- SeaSondeR:::seasonder_defaultSpecsFilePath("CS")
cs_obj <- seasonder_readSeaSondeCSFile(
  system.file("css_data/CSS_TORA_24_04_04_0700.cs", package = "SeaSondeR"),
  specs_path
)
# Define Doppler bin indices (from 1 to number of bins)
bins <- seq_len(as.integer(seasonder_getnDopplerCells(cs_obj)))
# Retrieve normalized Doppler frequencies for the specified bins (internal function)
normalized_freqs <- SeaSondeR:::seasonder_Bins2NormalizedDopplerFreq(cs_obj, bins)
head(normalized_freqs)
#> [1] -2.868662 -2.863048 -2.857435 -2.851821 -2.846207 -2.840593