Convert MUSIC Bearings to Geographic Bearings

This function converts MUSIC bearings (relative to the antenna) into geographic bearings using the antenna's bearing information from a SeaSondeRAPM object.

Usage

seasonder_MUSICBearing2GeographicalBearing(bearings, seasonder_apm_object)

Arguments

bearings

A list of numeric vectors containing MUSIC bearings in degrees. Each vector corresponds to a set of bearings relative to the antenna.

seasonder_apm_object

A SeaSondeRAPM object containing the antenna's metadata, including the antenna's bearing.

Value

A list of numeric vectors containing the geographic bearings in degrees. Each vector corresponds to a set of geographic bearings derived from the input.

Details

The geographic bearing is calculated by:

Multiplying the MUSIC bearings by -1 to invert their direction.

Adjusting the angles to the range [0, 360) using modulo 360.

Adding the antenna bearing to each value and wrapping the result to the range [0, 360) again using modulo 360.

The formula for each bearing is: [ geo\_bearing = ((-1 music\_bearing 360) + antenna\_bearing) 360 ]

See also

Examples

if (FALSE) { # \dontrun{
# Example conversion
music_bearings <- list(c(45, 90, 135), c(270, 315, 360))
antenna_apm <- seasonder_createSeaSondeRAPM()  # Assuming a valid SeaSondeRAPM object
geo_bearings <- seasonder_MUSICBearing2GeographicalBearing(music_bearings, antenna_apm)
print(geo_bearings)
} # }