Calculate the MUSIC Covariance Matrix for a Given Cell Range and Doppler Bin

This function computes the MUltiple SIgnal Classification (MUSIC) covariance matrix for a specific cell range and Doppler bin from SeaSonde Cross Spectra (CS) data. The MUSIC algorithm is used in direction finding and spectral estimation.

Usage

seasonder_MUSICComputeCov(seasonder_cs_object)

Arguments

seasonder_cs_obj

A SeaSondeRCS object containing the cross spectra data.

cell_range

A numeric vector specifying the range cells for which to calculate the covariance matrix.

doppler_bin

An integer specifying the Doppler bin for which the covariance matrix is calculated.

Value

A 3x3 complex matrix representing the MUSIC covariance matrix for the specified cell range and Doppler bin. Each element \(C_{ij}\) of the matrix is calculated based on the auto-spectra (for diagonal elements) or cross-spectra (for off-diagonal elements) data. For diagonal elements \(i = j\), \(C_{ii}\) is obtained from the auto-spectra SSA{i}, and for off-diagonal elements \(i \neq j\), \(C_{ij}\) is obtained from the cross-spectra CSij, where i and j are indices of the matrix.

Details

The MUSIC algorithm is widely used for estimating the direction of arrival of signals and requires the computation of a covariance matrix from sensor data. In the context of SeaSonde radar data, this function utilizes the cross spectra (CS) and auto-spectra (SSA) data to construct the covariance matrix necessary for MUSIC analysis. The function iterates over a 3x3 matrix to fill in the values based on whether the indices of the matrix are equal (diagonal elements) or not (off-diagonal elements). Diagonal elements are derived from auto-spectra data SSA{i}, where i corresponds to the antenna number. Off-diagonal elements are calculated from cross-spectra data CSij, representing the cross-spectrum between antennas i and j.