Eigen Decomposition of the MUSIC Covariance Matrix

Performs the eigen decomposition of a MUSIC covariance matrix to obtain the eigenvalues and eigenvectors. This decomposition is a critical step in the MUSIC algorithm for spectral estimation and direction finding, as it enables the identification of the signal and noise subspaces.

Usage

seasonder_MUSICCovDecomposition(seasonder_cs_object)

Arguments

seasonder_cs_object

A SeaSondeRCS object containing the covariance matrices derived from cross-spectra data.

Value

An updated SeaSondeRCS object where each Doppler cell includes the eigenvalues and eigenvectors of its covariance matrix. The eigenvalues are sorted in descending order, and the eigenvectors are aligned accordingly. The updates include:

eigen$values: A numeric vector containing the sorted eigenvalues for each Doppler cell.

eigen$vectors: A 3x3 matrix of the corresponding eigenvectors for each Doppler cell, aligned with the eigenvalues.

Details

The covariance matrix represents one Doppler cell of the averaged cross-spectra of three received signals. This matrix captures the summation of signals from all bearings (plus noise) received by the antennas. To estimate the direction of arrival (DOA), the covariance matrix is subjected to eigenvalue decomposition (diagonalization) to estimate the signal and noise subspaces.

In practical HF radar systems, there are two primary sources of noise:

System (thermal) noise: Generated by the receiving equipment and assumed to be uncorrelated between antennas.

Spatial noise field: Includes wind-wave noise and current noise, modeled as Gaussian, which introduces correlation.

The eigenvalue decomposition produces:

Three eigenvalues, ordered from largest to smallest.

Three corresponding eigenvectors forming a 3-dimensional orthonormal basis.

Based on the largest eigenvalues:

If there is one signal present, the first eigenvector defines a 1-dimensional signal subspace, and the remaining eigenvectors represent a 2-dimensional noise subspace.

If two signals are present, the first two eigenvectors form a 2-dimensional signal subspace, while the remaining eigenvector represents a 1-dimensional noise subspace.

The signal and noise subspaces are orthogonal. This decomposition facilitates identifying the signal's direction by finding the antenna manifold that best fits the signal subspace.

References

Paolo, T. de, Cook, T. & Terrill, E. Properties of HF RADAR Compact Antenna Arrays and Their Effect on the MUSIC Algorithm. OCEANS 2007 1–10 (2007) doi:10.1109/oceans.2007.4449265.

See also

Examples

if (FALSE) { # \dontrun{
# Assuming `cs_object` is a valid SeaSondeRCS object
cs_object <- seasonder_MUSICCovDecomposition(cs_object)
} # }