Compute First Order Regions (FOR) Based on Selected Method

This function processes a SeaSondeRCS object to compute the First Order Regions (FOR) using the specified method. It allows the user to configure the processing method and parameters dynamically.

Usage

seasonder_computeFORs(seasonder_cs_obj, method = NULL, FOR_control = NULL)

Arguments

seasonder_cs_obj

A SeaSondeRCS object containing the spectral data and FOR parameters.

method

Optional; A character string specifying the method to be used for FOR computation. Defaults to NULL, in which case the method stored in the object is used. Currently supported method: "SeaSonde".

FOR_control

Optional; A list of parameters for configuring the FOR computation process. Defaults to NULL, in which case the parameters already stored in the object are used.

Value

The updated SeaSondeRCS object with the computed First Order Regions (FOR).

Details

Steps:

Set Method:

If a method is provided, it updates the SeaSondeRCS object with the specified method.

Retrieve Method:

If no method is specified, the function retrieves the method stored in the object.

Set Parameters:

If FOR_control is provided, the function updates the object's FOR parameters.

Method Execution:

Based on the selected method, the corresponding processing function is called. Currently, the only supported method is "SeaSonde", which calls seasonder_computeFORsSeaSondeMethod.

Use Case: This function provides a flexible interface for computing FORs, allowing users to dynamically select methods and configure parameters without modifying the internal object structure.

See also

Examples

if (FALSE) { # \dontrun{
# Compute FORs using the SeaSonde method
cs_obj <- seasonder_computeFORs(cs_obj, method = "SeaSonde")

# Compute FORs with custom parameters
FOR_control <- list(currmax = 2, noisefact = 4)
cs_obj <- seasonder_computeFORs(cs_obj, FOR_control = FOR_control)
} # }