Create a SeaSondeRAPM Object

This function creates a SeaSondeRAPM object to store antenna pattern calibration data.

Usage

seasonder_createSeaSondeRAPM(
  calibration_matrix = matrix(complex(real = NA_real_, imaginary = NA_real_), nrow = 3,
    ncol = 0),
  ...
)

Arguments

calibration_matrix

A 3 x b complex matrix, where b is the number of bearings for calibration.

...

Additional named attributes that will be passed to seasonder_initializeAttributesSeaSondeRAPM.

Value

A SeaSondeRAPM object containing a complex matrix with class attribute 'SeaSondeRAPM' and additional attributes for metadata. Row names are set "A13", "A23" and "A33" and column names are set to be the values in BEAR.

Details

The function performs the following operations:

Validates the calibration_matrix with seasonder_validateCalibrationMatrixSeaSondeRAPM.

Initializes all other attributes either with default or user-provided values.

Merges the initialized attributes into calibration_matrix.

Sets the object's class to 'SeaSondeRAPM'.

For more details on the attributes, see seasonder_initializeAttributesSeaSondeRAPM.

See also