Create a SeaSondeRCS object

This generic function creates a SeaSondeRCS object either from a file path or directly from a list containing header and data. When x is a character string, the function determines the file type (either "CS" or "CSSY") by analyzing the spectra file and reads it using the appropriate reading function. If specs_path is not provided (or set to rlang::zap()), the default YAML specifications path corresponding to the detected file type is used.

Usage

seasonder_createSeaSondeRCS(x, specs_path = NULL, ...)

Arguments

x

Either a character string specifying the path to the SeaSonde CS file or a list containing header and data.

specs_path

A character string specifying the path to the YAML specifications for the CS file. Used only if x is a character string.

...

Additional parameters passed to the underlying functions.

endian

A character string indicating the byte order. Options are "big" (default) or "little".

Value

A SeaSondeRCS object.

Details

For character inputs, the function first checks if the specified file exists. It then determines the file type using seasonder_find_spectra_file_type. If the specs_path parameter is not provided or is set to rlang::zap(), the default specifications file path is obtained using seasonder_defaultSpecsFilePath based on the detected file type. The file is then read using the appropriate reading function:

seasonder_readSeaSondeCSFile for CS files.

seasonder_readSeaSondeRCSSYFile for CSSY files.

For list inputs, the SeaSondeRCS object is created directly from the provided header and data. Additionally, a processing step is appended to the object using seasonder_setSeaSondeRCS_ProcessingSteps with a creation step text that indicates the source.

See also

Examples

if (FALSE) { # \dontrun{
  # Creating a SeaSondeRCS object from a list
  cs_list <- list(header = my_header, data = my_data)
  rcs_object <- seasonder_createSeaSondeRCS(cs_list)

  # Creating a SeaSondeRCS object from a file path using default YAML specifications
  rcs_object <- seasonder_createSeaSondeRCS("path/to/cs_file.cs")

  # Creating a SeaSondeRCS object from a file path with a specified YAML specifications file
  rcs_object <- seasonder_createSeaSondeRCS("path/to/cs_file.cs", specs_path = "path/to/specs.yaml")
} # }