Default Options for the MUSIC Algorithm

This function returns a list of default options used in the MUSIC algorithm.

Usage

seasonder_defaultMUSICOptions()

Value

A list containing the default options for the MUSIC algorithm.

Details

The returned list includes:

PPMIN: Lower threshold value (default is NULL).

PWMAX: Upper threshold value (default is NULL).

smoothNoiseLevel: Logical flag indicating whether the noise level should be smoothed (FALSE by default).

doppler_interpolation: Doppler interpolation factor (default is 2).

MUSIC_parameters: A numeric vector of default parameters for the MUSIC algorithm, retrieved from seasonder_defaultMUSIC_parameters().

discard_low_SNR: Logical flag to discard solutions with low signal-to-noise ratio (TRUE by default).

discard_no_solution: Logical flag to discard cases with no solution (TRUE by default).

Examples

# Retrieve the default options for the MUSIC algorithm
opts <- seasonder_defaultMUSICOptions()
print(opts)
#> $PPMIN
#> NULL
#> 
#> $PWMAX
#> NULL
#> 
#> $smoothNoiseLevel
#> [1] FALSE
#> 
#> $doppler_interpolation
#> [1] 2
#> 
#> $MUSIC_parameters
#> [1] 40 20  2 20
#> 
#> $discard_low_SNR
#> [1] TRUE
#> 
#> $discard_no_solution
#> [1] TRUE
#>