Estimate Reference Noise Limits in Normalized Doppler Frequency

This function estimates the reference noise limits for normalized Doppler frequencies in a SeaSondeR cross-spectral object. These limits are used to define the frequency range where the noise level is assessed for first-order region (FOR) detection.

Usage

seasonder_estimateReferenceNoiseNormalizedLimits(seasonder_cs_obj)

Arguments

seasonder_cs_obj

A SeaSondeRCS object containing Doppler frequency metadata.

Value

A numeric vector of length two, representing the lower and upper reference noise limits in normalized Doppler frequency.

Details

The function computes the noise limits based on the maximum normalized Doppler frequency from seasonder_getDopplerBinsFrequency. The limits are scaled using fixed factors to set a reference range for noise analysis:

The lower bound is 56.5% of the maximum normalized Doppler frequency.

The upper bound is 100% of the maximum normalized Doppler frequency.

The lower bound value of 56.5% was determined through an empirical process involving multiple sample spectra. The procedure began with an initial lower bound set at 100%, and in each iteration, this value was decreased by 0.5%. The noise floor was calculated repeatedly and compared with the temporal NoiseFloor.txt file generated by the AnalyseSpectra Tool in Radial Suite R8. The final lower bound percentage was established as the value that closely matched the corresponding noise floor in the NoiseFloor.txt file. After this, several percentages were averaged and rounded to the nearest 0.5%.

The function is used in first-order detection processes where noise reference levels are required for signal-to-noise ratio (SNR) calculations.

See also

Examples

if (FALSE) { # \dontrun{
# Estimate reference noise limits for a SeaSondeRCS object
noise_limits <- seasonder_estimateReferenceNoiseNormalizedLimits(cs_obj)
print(noise_limits)
} # }