Retrieve First Order Region (FOR) Parameters

This function retrieves the First Order Region (FOR) parameters associated with a SeaSondeR cross-spectral object. If no FOR parameters are found in the object's attributes, it initializes them using seasonder_validateFOR_parameters.

Usage

seasonder_getFOR_parameters(seasonder_cs_object)

Arguments

seasonder_cs_object

A SeaSondeRCS object containing FOR-related metadata.

Value

A named list containing the validated FOR parameters.

Details

The function extracts the FOR parameters stored within the object. If the parameters are missing, the function initializes them using seasonder_validateFOR_parameters and assigns default values where necessary.

FOR Parameters:

nsm: Doppler smoothing factor.

fdown: Peak power dropoff threshold.

flim: Null below peak power threshold.

noisefact: Signal-to-noise threshold.

currmax: Maximum velocity allowed.

reject_distant_bragg: Flag to reject distant Bragg signals.

reject_noise_ionospheric: Flag to reject ionospheric noise contamination.

reject_noise_ionospheric_threshold: Threshold (in dB) for rejecting noise-affected Bragg signals.

reference_noise_normalized_limits: Estimated limits for reference noise in normalized Doppler frequency.

See also

Examples

# Minimal example for seasonder_getFOR_parameters
cs_file <- system.file("css_data/CSS_TORA_24_04_04_0700.cs", package = "SeaSondeR")
apm_file <- system.file("css_data/MeasPattern.txt", package = "SeaSondeR")
apm_obj <- seasonder_readSeaSondeRAPMFile(apm_file)
#> seasonder_createSeaSondeRAPM:  APM object created successfully.
cs_obj <- seasonder_createSeaSondeRCS(cs_file, seasonder_apm_object = apm_obj)
#> new_SeaSondeRCS:  SeaSondeRCS object created successfully.
for_params <- SeaSondeR:::seasonder_getFOR_parameters(cs_obj)
print(for_params)
#> $nsm
#> [1] 2
#> 
#> $reference_noise_normalized_limits
#> [1] 2.586848 2.874276
#> 
#> $fdown
#> [1] 10
#> 
#> $flim
#> [1] 100
#> 
#> $noisefact
#> [1] 3.981072
#> 
#> $currmax
#> [1] 2
#> 
#> $reject_distant_bragg
#> [1] TRUE
#> 
#> $reject_noise_ionospheric
#> [1] TRUE
#> 
#> $reject_noise_ionospheric_threshold
#> [1] 0
#>