Retrieve First Order Region (FOR) Configuration from a SeaSondeRCS Object

This function extracts the configuration related to the First Order Region (FOR) from a SeaSondeRCS object. It returns a list containing the FOR parameters and the noise level assigned to the object.

Usage

seasonder_getSeaSondeRCS_FORConfig(seasonder_cs_object)

Arguments

seasonder_cs_object

A SeaSondeRCS object containing FOR-related metadata.

Value

A list with two components:

FOR_parameters: A list of parameters used for FOR processing.

NoiseLevel: The noise level values retrieved from the object.

Details

The FOR configuration is composed of parameters that define the first order region and the noise level used during FOR processing. This function aggregates these components by calling seasonder_getFOR_parameters() and seasonder_getSeaSondeRCS_NoiseLevel().

Examples

# Create a minimal SeaSondeRCS object
cs_file <- system.file("css_data/CSS_TORA_24_04_04_0700.cs", package = "SeaSondeR")
apm_file <- system.file("css_data/MeasPattern.txt", package = "SeaSondeR")
apm_obj <- seasonder_readSeaSondeRAPMFile(apm_file)
#> seasonder_createSeaSondeRAPM:  APM object created successfully.
cs_obj <- seasonder_createSeaSondeRCS(cs_file, seasonder_apm_object = apm_obj)
#> new_SeaSondeRCS:  SeaSondeRCS object created successfully.
config <- seasonder_getSeaSondeRCS_FORConfig(cs_obj)
print(config)
#> $FOR_parameters
#> $FOR_parameters$nsm
#> [1] 2
#> 
#> $FOR_parameters$reference_noise_normalized_limits
#> [1] 2.586848 2.874276
#> 
#> $FOR_parameters$fdown
#> [1] 10
#> 
#> $FOR_parameters$flim
#> [1] 100
#> 
#> $FOR_parameters$noisefact
#> [1] 3.981072
#> 
#> $FOR_parameters$currmax
#> [1] 2
#> 
#> $FOR_parameters$reject_distant_bragg
#> [1] TRUE
#> 
#> $FOR_parameters$reject_noise_ionospheric
#> [1] TRUE
#> 
#> $FOR_parameters$reject_noise_ionospheric_threshold
#> [1] 0
#> 
#> 
#> $NoiseLevel
#> range_001 range_002 range_003 range_004 range_005 range_006 range_007 range_008 
#> -135.2474 -134.8318 -135.0454 -135.0340 -134.9676 -134.7360 -134.8228 -134.8941 
#> range_009 range_010 range_011 range_012 range_013 range_014 range_015 range_016 
#> -134.8782 -135.0698 -135.2970 -135.1301 -134.8897 -134.8974 -134.8161 -134.7336 
#> range_017 range_018 range_019 range_020 range_021 range_022 range_023 range_024 
#> -134.9820 -134.5720 -134.0027 -134.4659 -134.9164 -135.5818 -135.4132 -135.2606 
#> range_025 range_026 range_027 range_028 range_029 range_030 range_031 range_032 
#> -135.4979 -135.5128 -135.6235 -135.3909 -135.3739 -135.2351 -135.7927 -135.7334 
#> range_033 range_034 range_035 range_036 range_037 range_038 range_039 range_040 
#> -135.6103 -135.7485 -136.0722 -136.1765 -135.4616 -135.9233 -136.4067 -136.0290 
#> range_041 range_042 range_043 range_044 range_045 range_046 range_047 range_048 
#> -135.9551 -136.3613 -136.1284 -136.6700 -136.8593 -136.8549 -136.5435 -136.2426 
#> range_049 range_050 range_051 range_052 range_053 range_054 range_055 range_056 
#> -136.1255 -136.2856 -136.2771 -136.7757 -136.7783 -136.6267 -136.8762 -136.7339 
#> range_057 range_058 range_059 range_060 range_061 range_062 range_063 
#> -136.5483 -136.7173 -136.6149 -136.7406 -136.8187 -136.9787 -137.3383 
#>