Retrieve MUSIC Data from a SeaSondeRCS Object

This function extracts the MUSIC data structure from a SeaSondeRCS object.

Usage

seasonder_getSeaSondeRCS_MUSIC(seasonder_cs_object)

Arguments

seasonder_cs_object

A SeaSondeRCS object containing MUSIC data as an attribute.

Value

The MUSIC data structure, typically a data frame or tibble with MUSIC results.

Details

If the MUSIC data does not exist in the object, the function initializes it via seasonder_initSeaSondeRCS_MUSIC().

Examples

# Minimal example for seasonder_getMUSIC
cs_file <- system.file("css_data/CSS_TORA_24_04_04_0700.cs", package = "SeaSondeR")
apm_file <- system.file("css_data/MeasPattern.txt", package = "SeaSondeR")
apm_obj <- seasonder_readSeaSondeRAPMFile(apm_file)
#> seasonder_createSeaSondeRAPM:  APM object created successfully.
cs_obj <- seasonder_createSeaSondeRCS(cs_file, seasonder_apm_object = apm_obj)
#> new_SeaSondeRCS:  SeaSondeRCS object created successfully.
music_data <- SeaSondeR:::seasonder_getMUSIC(cs_obj)
print(music_data)
#> # A tibble: 64,512 × 18
#>    range_cell doppler_bin range  freq radial_v cov            MA1S  MA2S  MA3S
#>         <int>       <int> <dbl> <dbl>    <dbl> <list>        <dbl> <dbl> <dbl>
#>  1          1           1 0.187 -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#>  2          2           1 0.374 -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#>  3          3           1 0.561 -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#>  4          4           1 0.748 -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#>  5          5           1 0.935 -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#>  6          6           1 1.12  -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#>  7          7           1 1.31  -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#>  8          8           1 1.50  -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#>  9          9           1 1.68  -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#> 10         10           1 1.87  -2.00    -4.19 <cpl [3 × 3]>    NA    NA    NA
#> # ℹ 64,502 more rows
#> # ℹ 9 more variables: eigen <list>, projections <list>, DOA_solutions <list>,
#> #   eigen_values_ratio <dbl>, P1_check <lgl>, retained_solution <chr>,
#> #   DOA <list>, lonlat <list>, VFLG <dbl>