Convert an integer to raw bytes using a 64-bit representation

This function converts an integer to a raw byte representation using a 64-bit (8-byte) format. It leverages the bit64 package to handle the 64-bit integer representation and conversion.

Usage

seasonder_int_to_raw(x)

Arguments

x

An integer to be converted to raw bytes.

Value

A raw vector representing the 64-bit format of the provided integer.

Details

The function follows these steps:

Convert the integer to a 64-bit format using bit64::as.integer64.

Convert the 64-bit integer to a bit string.

Split the bit string into individual bits.

Reorder the bits into groups of 8, reversing the order within each group.

Convert the reordered bits back to raw bytes.

Examples

if (FALSE) { # \dontrun{
int_val <- 1234567890
raw_val <- seasonder_int_to_raw(int_val)
cat(rawToChar(raw_val))
} # }