Read the SeaSonde CS File Header

This function reads and processes the header of a SeaSonde CS file. It initially reads the general header (Version 1) to determine the file version. Subsequent headers are processed based on the file version.

Usage

seasonder_readSeaSondeCSFileHeader(specs, connection, endian = "big")

Arguments

specs

List of header specifications for each version.

connection

The file connection.

endian

Character string indicating the byte order, either "big" (default) or "little".

Value

A combined list of all processed headers up to the file version.

See also

Examples

# Example: read header from a sample CS file
specs_path <- SeaSondeR:::seasonder_defaultSpecsFilePath("CS")
specs <- seasonder_readYAMLSpecs(specs_path, "header")
con <- file(system.file("css_data/CSS_TORA_24_04_04_0700.cs", package = "SeaSondeR"), "rb")
header <- SeaSondeR:::seasonder_readSeaSondeCSFileHeader(specs, con, endian = "big")
print(header)
#> $nCsFileVersion
#> [1] 6
#> 
#> $nDateTime
#> [1] "2024-04-04 07:00:00 UTC"
#> 
#> $nV1Extent
#> [1] 1319
#> 
#> $nCsKind
#> [1] 2
#> 
#> $nV2Extent
#> [1] 1313
#> 
#> $nSiteCodeName
#> [1] "TORA"
#> 
#> $nV3Extent
#> [1] 1305
#> 
#> $nRangeCells
#> [1] 63
#> 
#> $nDopplerCells
#> [1] 1024
#> 
#> $nFirstRangeCell
#> [1] 1
#> 
#> $nCoverMinutes
#> [1] 15
#> 
#> $bDeletedSource
#> [1] 0
#> 
#> $bOverrideSrcInfo
#> [1] 0
#> 
#> $fStartFreqMHz
#> [1] 46.90071
#> 
#> $fRepFreqHz
#> [1] 4
#> 
#> $fBandwidthKHz
#> [1] 801.4276
#> 
#> $bSweepUp
#> [1] 0
#> 
#> $fRangeCellDistKm
#> [1] 0.1870365
#> 
#> $nV4Extent
#> [1] 1257
#> 
#> $CenterFreq
#> [1] 46.5
#> 
#> $CellsDistKm
#>  [1]  0.1870365  0.3740731  0.5611096  0.7481461  0.9351826  1.1222192
#>  [7]  1.3092557  1.4962922  1.6833288  1.8703653  2.0574018  2.2444384
#> [13]  2.4314749  2.6185114  2.8055479  2.9925845  3.1796210  3.3666575
#> [19]  3.5536941  3.7407306  3.9277671  4.1148036  4.3018402  4.4888767
#> [25]  4.6759132  4.8629498  5.0499863  5.2370228  5.4240593  5.6110959
#> [31]  5.7981324  5.9851689  6.1722055  6.3592420  6.5462785  6.7333151
#> [37]  6.9203516  7.1073881  7.2944246  7.4814612  7.6684977  7.8555342
#> [43]  8.0425708  8.2296073  8.4166438  8.6036803  8.7907169  8.9777534
#> [49]  9.1647899  9.3518265  9.5388630  9.7258995  9.9129360 10.0999726
#> [55] 10.2870091 10.4740456 10.6610822 10.8481187 11.0351552 11.2221918
#> [61] 11.4092283 11.5962648 11.7833013
#> 
#> $nOutputInterval
#> [1] 4
#> 
#> $nCreateTypeCode
#> [1] "SSAQ"
#> 
#> $nCreatorVersion
#> [1] "11.9"
#> 
#> $nActiveChannels
#> [1] 3
#> 
#> $nSpectraChannels
#> [1] 3
#> 
#> $nActiveChanBits
#> [1] 7
#> 
#> $nV5Extent
#> [1] 1229
#> 
#> $ActiveChannels
#> [1] 1 2 3
#> 
#> $nCS6ByteSize
#> [1] 1225
#> 
#> $TIME
#> $TIME$nTimeMark
#> [1] start
#> Levels: start center time end time
#> 
#> $TIME$nYear
#> [1] 2024
#> 
#> $TIME$nMonth
#> [1] 4
#> 
#> $TIME$nDay
#> [1] 4
#> 
#> $TIME$nHour
#> [1] 7
#> 
#> $TIME$nMinute
#> [1] 0
#> 
#> $TIME$fSeconds
#> [1] 0
#> 
#> $TIME$fCoverageSeconds
#> [1] 3.506412e-310
#> 
#> $TIME$fHoursFromUTC
#> [1] 0
#> 
#> 
#> $ZONE
#> $ZONE$szTimeZone
#> [1] "Atlantic/Reykjavik"
#> 
#> 
#> $LOCA
#> $LOCA$fLatitude
#> [1] 42.20127
#> 
#> $LOCA$fLongitude
#> [1] -8.801883
#> 
#> $LOCA$fAltitudeMeters
#> [1] 0
#> 
#> 
#> $RCVI
#> $RCVI$nReceiverModel
#> [1] Unknown
#> 6 Levels: Unknown Awg3/Rcvr2 Chassis AC Awg3/Rcvr2 Chassis DC ... Awg4 Chassis DC
#> 
#> $RCVI$nRxAntennaModel
#> [1] Unknown
#> Levels: Unknown Box Loops 2 3 Dome Loops TR Dome Loops
#> 
#> $RCVI$fReferenceGainDB
#> [1] 34.2
#> 
#> $RCVI$szFirmware
#> [1] ""
#> 
#> 
#> $GLRM
#> $GLRM$nMethod
#> [1] 2
#> 
#> $GLRM$nVersion
#> [1] 1
#> 
#> $GLRM$nPointsRemoved
#> [1] 0
#> 
#> $GLRM$nTimesRemoved
#> [1] 0
#> 
#> $GLRM$nSegmentsRemoved
#> [1] 0
#> 
#> $GLRM$fPointPowerThreshold
#> [1] 15
#> 
#> $GLRM$fRangePowerThreshold
#> [1] 10
#> 
#> $GLRM$fRangeBinThreshold
#> [1] 0.5
#> 
#> $GLRM$bRemoveDC
#> [1] 0
#> 
#> $GLRM$nReceiverModel
#> factor()
#> Levels: Off Point Range Range&Point SubDCOnly
#> 
#> 
#> $FOLS
#> $FOLS$nNegBraggLeftIndex
#> $FOLS$nNegBraggLeftIndex$loop
#> [1] "nRangeCells"
#> 
#> $FOLS$nNegBraggLeftIndex$data
#>  [1]   0 334 335 321 317 324 323 323 316 313 314 312 325 312 312 314 315 316 317
#> [20] 317 316 312 311 311 313 312 308 305 308 305 305 304 310 305 301 299 300 304
#> [39] 296 299 297 297 300 301 302 302 300 300   0   0   0   0   0   0   0   0   0
#> [58]   0   0   0   0   0   0
#> 
#> 
#> $FOLS$nNegBraggRightIndex
#> $FOLS$nNegBraggRightIndex$loop
#> [1] "nRangeCells"
#> 
#> $FOLS$nNegBraggRightIndex$data
#>  [1]   0 333 340 344 349 351 349 349 352 353 351 353 349 348 347 346 346 345 342
#> [20] 334 348 337 337 341 340 340 349 343 347 347 348 336 337 356 353 335 362 356
#> [39] 359 341 352 350 347 343 343 342 350 351   0   0   0   0   0   0   0   0   0
#> [58]   0   0   0   0   0   0
#> 
#> 
#> $FOLS$nPosBraggLeftIndex
#> $FOLS$nPosBraggLeftIndex$loop
#> [1] "nRangeCells"
#> 
#> $FOLS$nPosBraggLeftIndex$data
#>  [1]   0 689 689 673 668 669 664 664 666 666 665 667 667 665 666 667 667 665 665
#> [20] 664 665 665 663 666 667 668 667 663 662 662 662 663 664 662 660 665 663 658
#> [39] 609 668 609 609 609 609 609 624 657 660   0   0   0   0   0   0   0   0   0
#> [58]   0   0   0   0   0   0
#> 
#> 
#> $FOLS$nPosBraggRightIndex
#> $FOLS$nPosBraggRightIndex$loop
#> [1] "nRangeCells"
#> 
#> $FOLS$nPosBraggRightIndex$data
#>  [1]   0 688 688 684 698 706 704 707 684 681 682 684 684 709 692 705 707 690 690
#> [20] 687 687 699 693 690 695 696 692 693 694 694 692 691 703 705 691 693 702 695
#> [39] 696 695 664 697 654 653 672 692 679 680   0   0   0   0   0   0   0   0   0
#> [58]   0   0   0   0   0   0
#> 
#> 
#> 
close(con)