Read SeaSonde CS File Header V6

This function reads the header of a SeaSonde CS File Version 6. It sequentially reads blocks based on the provided specifications and returns the read data.

Usage

seasonder_readSeaSondeCSFileHeaderV6(
  specs,
  connection,
  endian = "big",
  prev_data = NULL
)

Arguments

specs

A list of specifications for reading the file header. It should contain three main elements: nCS6ByteSize, block_spec, and blocks, each containing further specifications for reading various parts of the header.

connection

A connection object to the SeaSonde CS file.

endian

The byte order for reading the file. Default is "big".

prev_data

Previous data, if any, that might affect the current reading. Default is NULL.

Value

A list containing the read data, organized based on the block keys.

Condition Management

This function utilizes the rlang package to manage conditions and provide detailed and structured condition messages:

Condition Classes:

seasonder_v6_block_transformacion_skipped: Triggered when a transformation for a specific block is skipped.

seasonder_v6_transform_function_error: Triggered when there's an error while applying the transformation function for a V6 header block.

seasonder_v6_skip_block_error: Triggered when there's an error while skipping a block.

Condition Cases:

The following are the scenarios when errors or conditions are raised:

Transformation Failure: If there's a recognized block key and the transformation function associated with it fails.

Error in Transformation Function Application: If there's an error while applying the transformation function for a recognized V6 header block.

Error in Skipping Block: If there's an error while skipping a block when the block key is not recognized.

Restart Options:

The function provides the following restart option:

seasonder_v6_skip_transformation: This restart allows users to skip the transformation for a specific block and instead return the provided value.

Effects of Restart Options:

Using the seasonder_v6_skip_transformation restart:

The error message gets logged.

The transformation that caused the error gets skipped.

The provided value for that block is returned.

Proper error management ensures the integrity of the reading process and provides detailed feedback to users regarding issues and potential resolutions.

See also

Examples

# Read CS file header (Version 6 blocks) via wrapper
specs_path <- SeaSondeR:::seasonder_defaultSpecsFilePath("CS")
con <- file(
  system.file("css_data/CSS_TORA_24_04_04_0700.cs", package = "SeaSondeR"),
  "rb"
)
specs <- SeaSondeR:::seasonder_readYAMLSpecs(specs_path, c("header","V1"))
header_v1 <- SeaSondeR:::seasonder_readSeaSondeCSFileHeaderV1(
  specs,
  con,
  endian = "big"
)
specs <- SeaSondeR:::seasonder_readYAMLSpecs(specs_path, c("header","V2"))
header_v2 <- SeaSondeR:::seasonder_readSeaSondeCSFileHeaderV2(
  specs,
  con,
  endian = "big"
)
specs <- SeaSondeR:::seasonder_readYAMLSpecs(specs_path, c("header","V3"))
header_v3 <- SeaSondeR:::seasonder_readSeaSondeCSFileHeaderV3(
  specs,
  con,
  endian = "big"
)
specs <- SeaSondeR:::seasonder_readYAMLSpecs(specs_path, c("header","V4"))
header_v4 <- SeaSondeR:::seasonder_readSeaSondeCSFileHeaderV4(
  specs,
  con,
  endian = "big"
)
specs <- SeaSondeR:::seasonder_readYAMLSpecs(specs_path, c("header","V5"))
header_v5 <- SeaSondeR:::seasonder_readSeaSondeCSFileHeaderV5(
  specs,
  con,
  endian = "big"
)
specs <- SeaSondeR:::seasonder_readYAMLSpecs(specs_path, c("header","V6"))
prev_data <- purrr::reduce(list(header_v2,header_v3,header_v4,header_v5), \(pool, current_header){

  overlapping_keys <- intersect(names(pool), names(current_header))
  pool[overlapping_keys] <- current_header[overlapping_keys]

  
  new_keys <- setdiff(names(current_header), names(pool))
  pool <- c(pool, current_header[new_keys])
pool
}, .init = header_v1)
header_v6 <- SeaSondeR:::seasonder_readSeaSondeCSFileHeaderV6(
  specs,
  con,
  endian = "big",
prev_data = prev_data
)
str(header_v6)
#> List of 7
#>  $ nCS6ByteSize: num 1225
#>  $ TIME        :List of 9
#>   ..$ nTimeMark       : Factor w/ 3 levels "start","center time",..: 1
#>   ..$ nYear           : int 2024
#>   ..$ nMonth          : int 4
#>   ..$ nDay            : int 4
#>   ..$ nHour           : int 7
#>   ..$ nMinute         : int 0
#>   ..$ fSeconds        : num 0
#>   ..$ fCoverageSeconds: num 3.51e-310
#>   ..$ fHoursFromUTC   : num 0
#>  $ ZONE        :List of 1
#>   ..$ szTimeZone: chr "Atlantic/Reykjavik"
#>  $ LOCA        :List of 3
#>   ..$ fLatitude      : num 42.2
#>   ..$ fLongitude     : num -8.8
#>   ..$ fAltitudeMeters: num 0
#>  $ RCVI        :List of 4
#>   ..$ nReceiverModel  : Factor w/ 6 levels "Unknown","Awg3/Rcvr2 Chassis AC",..: 1
#>   ..$ nRxAntennaModel : Factor w/ 6 levels "Unknown","Box Loops",..: 1
#>   ..$ fReferenceGainDB: num 34.2
#>   ..$ szFirmware      : chr ""
#>  $ GLRM        :List of 10
#>   ..$ nMethod             : int 2
#>   ..$ nVersion            : int 1
#>   ..$ nPointsRemoved      : num 0
#>   ..$ nTimesRemoved       : num 0
#>   ..$ nSegmentsRemoved    : num 0
#>   ..$ fPointPowerThreshold: num 15
#>   ..$ fRangePowerThreshold: num 10
#>   ..$ fRangeBinThreshold  : num 0.5
#>   ..$ bRemoveDC           : int 0
#>   ..$ nReceiverModel      : Factor w/ 5 levels "Off","Point",..: 
#>  $ FOLS        :List of 4
#>   ..$ nNegBraggLeftIndex :List of 2
#>   .. ..$ loop: chr "nRangeCells"
#>   .. ..$ data: int [1:63] 0 334 335 321 317 324 323 323 316 313 ...
#>   ..$ nNegBraggRightIndex:List of 2
#>   .. ..$ loop: chr "nRangeCells"
#>   .. ..$ data: int [1:63] 0 333 340 344 349 351 349 349 352 353 ...
#>   ..$ nPosBraggLeftIndex :List of 2
#>   .. ..$ loop: chr "nRangeCells"
#>   .. ..$ data: int [1:63] 0 689 689 673 668 669 664 664 666 666 ...
#>   ..$ nPosBraggRightIndex:List of 2
#>   .. ..$ loop: chr "nRangeCells"
#>   .. ..$ data: int [1:63] 0 688 688 684 698 706 704 707 684 681 ...
close(con)