Read SeaSonde RCSSW File and Create SeaSondeRCS Object

This function reads a SeaSonde RCSSW file from a specified file path and parses its content into a SeaSondeRCS object. The file is processed by reading its header and body sections using CSSW specifications provided via a YAML file.

Usage

seasonder_readSeaSondeRCSSWFile(
  filepath,
  specs_path = seasonder_defaultSpecsFilePath("CSSW"),
  endian = "big"
)

Arguments

filepath

A character string specifying the path to the SeaSonde RCSSW file.

specs_path

A character string specifying the path to the YAML file containing CSSW specifications. Defaults to the output of seasonder_defaultSpecsFilePath("CSSW").

endian

A character string indicating the byte order used in the file. Defaults to "big".

Value

A SeaSondeRCS object containing the parsed header and data.

Details

The function executes the following steps:

Sets up error handling parameters specific to the function.

Retrieves YAML specifications for the key size block from the CSSW spec file.

Attempts to open the file in binary mode ("rb") with warnings suppressed.

Reads the file key and uses it to extract file specs.

Reads the header key, retrieves header specs, and parses the CSSW header.

Converts the CSSW header into a valid SeaSondeRCS header.

Reads the body key, retrieves body specs, and parses the CSSW body.

Transforms the CSSW body into a SeaSondeRCS data structure.

Combines the header and data into a SeaSondeRCS object.

Examples

# \donttest{
  # Example:
  cs_obj <- SeaSondeR:::seasonder_readSeaSondeRCSSWFile(
    system.file("css_data/CSS_TORA_2024_04_04_070000.csr", package = "SeaSondeR")
  )
#> new_SeaSondeRCS:  SeaSondeRCS object created successfully.
  str(cs_obj)
#> List of 2
#>  $ header:List of 37
#>   ..$ nCsFileVersion  : int 6
#>   ..$ nDateTime       : POSIXct[1:1], format: "2024-04-04 07:00:00"
#>   ..$ nV1Extent       : int 1319
#>   ..$ nCsKind         : int 2
#>   ..$ nV2Extent       : int 1313
#>   ..$ nSiteCodeName   : chr "TORA"
#>   ..$ nV3Extent       : int 1305
#>   ..$ nRangeCells     : int 63
#>   ..$ nDopplerCells   : int 1024
#>   ..$ nFirstRangeCell : int 1
#>   ..$ nCoverMinutes   : int 15
#>   ..$ bDeletedSource  : int 0
#>   ..$ bOverrideSrcInfo: int 0
#>   ..$ fStartFreqMHz   : num 46.9
#>   ..$ fRepFreqHz      : num 4
#>   ..$ fBandwidthKHz   : num 801
#>   ..$ bSweepUp        : int 0
#>   ..$ fRangeCellDistKm: num 0.187
#>   ..$ nV4Extent       : int 1257
#>   ..$ CenterFreq      : num 46.5
#>   ..$ CellsDistKm     : num [1:63] 0.187 0.374 0.561 0.748 0.935 ...
#>   ..$ nOutputInterval : int 4
#>   ..$ nCreateTypeCode : chr "SSAQ"
#>   ..$ nCreatorVersion : chr "11.9"
#>   ..$ nActiveChannels : int 3
#>   ..$ nSpectraChannels: int 3
#>   ..$ nActiveChanBits : num 7
#>   ..$ nV5Extent       : int 1229
#>   ..$ ActiveChannels  : int [1:3] 1 2 3
#>   ..$ nCS6ByteSize    : num 1225
#>   ..$ TIME            :List of 9
#>   .. ..$ nTimeMark       : Factor w/ 3 levels "start","center time",..: 1
#>   .. ..$ nYear           : int 2024
#>   .. ..$ nMonth          : int 4
#>   .. ..$ nDay            : int 4
#>   .. ..$ nHour           : int 7
#>   .. ..$ nMinute         : int 0
#>   .. ..$ fSeconds        : num 0
#>   .. ..$ fCoverageSeconds: num 3.51e-310
#>   .. ..$ fHoursFromUTC   : num 0
#>   ..$ ZONE            :List of 1
#>   .. ..$ szTimeZone: chr "Atlantic/Reykjavik"
#>   ..$ LOCA            :List of 3
#>   .. ..$ fLatitude      : num 42.2
#>   .. ..$ fLongitude     : num -8.8
#>   .. ..$ fAltitudeMeters: num 0
#>   ..$ RCVI            :List of 4
#>   .. ..$ nReceiverModel  : Factor w/ 6 levels "Unknown","Awg3/Rcvr2 Chassis AC",..: 1
#>   .. ..$ nRxAntennaModel : Factor w/ 6 levels "Unknown","Box Loops",..: 1
#>   .. ..$ fReferenceGainDB: num 34.2
#>   .. ..$ szFirmware      : chr ""
#>   ..$ GLRM            :List of 10
#>   .. ..$ nMethod             : int 2
#>   .. ..$ nVersion            : int 1
#>   .. ..$ nPointsRemoved      : num 0
#>   .. ..$ nTimesRemoved       : num 0
#>   .. ..$ nSegmentsRemoved    : num 0
#>   .. ..$ fPointPowerThreshold: num 15
#>   .. ..$ fRangePowerThreshold: num 10
#>   .. ..$ fRangeBinThreshold  : num 0.5
#>   .. ..$ bRemoveDC           : int 0
#>   .. ..$ nReceiverModel      : Factor w/ 5 levels "Off","Point",..: 
#>   ..$ FOLS            :List of 4
#>   .. ..$ nNegBraggLeftIndex :List of 2
#>   .. .. ..$ loop: chr "nRangeCells"
#>   .. .. ..$ data: int [1:63] 0 334 335 321 317 324 323 323 316 313 ...
#>   .. ..$ nNegBraggRightIndex:List of 2
#>   .. .. ..$ loop: chr "nRangeCells"
#>   .. .. ..$ data: int [1:63] 0 333 340 344 349 351 349 349 352 353 ...
#>   .. ..$ nPosBraggLeftIndex :List of 2
#>   .. .. ..$ loop: chr "nRangeCells"
#>   .. .. ..$ data: int [1:63] 0 689 689 673 668 669 664 664 666 666 ...
#>   .. ..$ nPosBraggRightIndex:List of 2
#>   .. .. ..$ loop: chr "nRangeCells"
#>   .. .. ..$ data: int [1:63] 0 688 688 684 698 706 704 707 684 681 ...
#>   ..$ header_csr      :List of 4
#>   .. ..$ sign:List of 7
#>   .. .. ..$ FileVersion: num 8.25e+08
#>   .. .. ..$ nFileType  : num 1.13e+09
#>   .. .. ..$ nOwner     : num 1.41e+09
#>   .. .. ..$ nUserFlags : num 0
#>   .. .. ..$ szFileName : chr "SeaSonde Shortened CrossSpectra"
#>   .. .. ..$ szOwnerName: chr "CODAR Ocean Sensors, Ltd"
#>   .. .. ..$ szComment  : chr ""
#>   .. ..$ srcn:List of 1
#>   .. .. ..$ SourceCSFilename: chr "CSS_TORA_24_04_04_0700.cs"
#>   .. ..$ mcda:List of 1
#>   .. .. ..$ FirstSweepTimestamp: num 3.8e+09
#>   .. ..$ dbrf:List of 1
#>   .. .. ..$ dBmReference: num 0
#>  $ data  :List of 7
#>   ..$ SSA1: num [1:63, 1:1024] 4.54e-11 3.38e-11 3.05e-11 3.80e-11 3.00e-11 ...
#>   ..$ SSA2: num [1:63, 1:1024] 3.47e-11 5.48e-11 8.77e-11 1.05e-10 2.90e-11 ...
#>   ..$ SSA3: num [1:63, 1:1024] -1.19e-10 -1.91e-10 -3.21e-10 -3.16e-10 -1.56e-10 ...
#>   ..$ CS12: cplx [1:63, 1:1024] -3.07e-11+6.09e-12i -3.10e-11+1.74e-11i -3.10e-11+6.69e-12i ...
#>   ..$ CS13: cplx [1:63, 1:1024] 5.08e-11+3.34e-11i 6.00e-11+2.59e-11i 5.35e-11+4.69e-11i ...
#>   ..$ CS23: cplx [1:63, 1:1024] -3.79e-11-4.54e-11i -5.51e-11-8.05e-11i -9.26e-11-1.30e-10i ...
#>   ..$ QC  : num [1:63, 1:1024] 1.26 1.26 1.26 1.26 1.26 ...
#>  - attr(*, "version")= num 1
#>  - attr(*, "ProcessingSteps")= chr [1:5] "2025-04-21 19:34:05.342992: FOR parameters set." "2025-04-21 19:34:05.353999: NoiseLevel for antenna 1 computed." "2025-04-21 19:34:05.369526: NoiseLevel for antenna 2 computed." "2025-04-21 19:34:05.380883: NoiseLevel for antenna 3 computed." ...
#>  - attr(*, "FOR_data")=List of 2
#>   ..$ FOR_parameters:List of 9
#>   .. ..$ nsm                               : num 2
#>   .. ..$ reference_noise_normalized_limits : num [1:2] 2.59 2.87
#>   .. ..$ fdown                             : num 10
#>   .. ..$ flim                              : num 100
#>   .. ..$ noisefact                         : num 3.98
#>   .. ..$ currmax                           : num 2
#>   .. ..$ reject_distant_bragg              : logi TRUE
#>   .. ..$ reject_noise_ionospheric          : logi TRUE
#>   .. ..$ reject_noise_ionospheric_threshold: num 0
#>   ..$ FOR           :List of 63
#>   .. ..$ range_001:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_002:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_003:List of 2
#>   .. .. ..$ negative_FOR: int [1:6] 336 337 338 339 340 341
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_004:List of 2
#>   .. .. ..$ negative_FOR: int [1:24] 322 323 324 325 326 327 328 329 330 331 ...
#>   .. .. ..$ positive_FOR: int [1:12] 674 675 676 677 678 679 680 681 682 683 ...
#>   .. ..$ range_005:List of 2
#>   .. .. ..$ negative_FOR: int [1:33] 318 319 320 321 322 323 324 325 326 327 ...
#>   .. .. ..$ positive_FOR: int [1:31] 669 670 671 672 673 674 675 676 677 678 ...
#>   .. ..$ range_006:List of 2
#>   .. .. ..$ negative_FOR: int [1:28] 325 326 327 328 329 330 331 332 333 334 ...
#>   .. .. ..$ positive_FOR: int [1:38] 670 671 672 673 674 675 676 677 678 679 ...
#>   .. ..$ range_007:List of 2
#>   .. .. ..$ negative_FOR: int [1:27] 324 325 326 327 328 329 330 331 332 333 ...
#>   .. .. ..$ positive_FOR: int [1:41] 665 666 667 668 669 670 671 672 673 674 ...
#>   .. ..$ range_008:List of 2
#>   .. .. ..$ negative_FOR: int [1:27] 324 325 326 327 328 329 330 331 332 333 ...
#>   .. .. ..$ positive_FOR: int [1:44] 665 666 667 668 669 670 671 672 673 674 ...
#>   .. ..$ range_009:List of 2
#>   .. .. ..$ negative_FOR: int [1:37] 317 318 319 320 321 322 323 324 325 326 ...
#>   .. .. ..$ positive_FOR: int [1:19] 667 668 669 670 671 672 673 674 675 676 ...
#>   .. ..$ range_010:List of 2
#>   .. .. ..$ negative_FOR: int [1:41] 314 315 316 317 318 319 320 321 322 323 ...
#>   .. .. ..$ positive_FOR: int [1:16] 667 668 669 670 671 672 673 674 675 676 ...
#>   .. ..$ range_011:List of 2
#>   .. .. ..$ negative_FOR: int [1:38] 315 316 317 318 319 320 321 322 323 324 ...
#>   .. .. ..$ positive_FOR: int [1:18] 666 667 668 669 670 671 672 673 674 675 ...
#>   .. ..$ range_012:List of 2
#>   .. .. ..$ negative_FOR: int [1:42] 313 314 315 316 317 318 319 320 321 322 ...
#>   .. .. ..$ positive_FOR: int [1:18] 668 669 670 671 672 673 674 675 676 677 ...
#>   .. ..$ range_013:List of 2
#>   .. .. ..$ negative_FOR: int [1:25] 326 327 328 329 330 331 332 333 334 335 ...
#>   .. .. ..$ positive_FOR: int [1:18] 668 669 670 671 672 673 674 675 676 677 ...
#>   .. ..$ range_014:List of 2
#>   .. .. ..$ negative_FOR: int [1:37] 313 314 315 316 317 318 319 320 321 322 ...
#>   .. .. ..$ positive_FOR: int [1:45] 666 667 668 669 670 671 672 673 674 675 ...
#>   .. ..$ range_015:List of 2
#>   .. .. ..$ negative_FOR: int [1:36] 313 314 315 316 317 318 319 320 321 322 ...
#>   .. .. ..$ positive_FOR: int [1:27] 667 668 669 670 671 672 673 674 675 676 ...
#>   .. ..$ range_016:List of 2
#>   .. .. ..$ negative_FOR: int [1:33] 315 316 317 318 319 320 321 322 323 324 ...
#>   .. .. ..$ positive_FOR: int [1:39] 668 669 670 671 672 673 674 675 676 677 ...
#>   .. ..$ range_017:List of 2
#>   .. .. ..$ negative_FOR: int [1:32] 316 317 318 319 320 321 322 323 324 325 ...
#>   .. .. ..$ positive_FOR: int [1:41] 668 669 670 671 672 673 674 675 676 677 ...
#>   .. ..$ range_018:List of 2
#>   .. .. ..$ negative_FOR: int [1:30] 317 318 319 320 321 322 323 324 325 326 ...
#>   .. .. ..$ positive_FOR: int [1:26] 666 667 668 669 670 671 672 673 674 675 ...
#>   .. ..$ range_019:List of 2
#>   .. .. ..$ negative_FOR: int [1:26] 318 319 320 321 322 323 324 325 326 327 ...
#>   .. .. ..$ positive_FOR: int [1:26] 666 667 668 669 670 671 672 673 674 675 ...
#>   .. ..$ range_020:List of 2
#>   .. .. ..$ negative_FOR: int [1:18] 318 319 320 321 322 323 324 325 326 327 ...
#>   .. .. ..$ positive_FOR: int [1:24] 665 666 667 668 669 670 671 672 673 674 ...
#>   .. ..$ range_021:List of 2
#>   .. .. ..$ negative_FOR: int [1:33] 317 318 319 320 321 322 323 324 325 326 ...
#>   .. .. ..$ positive_FOR: int [1:23] 666 667 668 669 670 671 672 673 674 675 ...
#>   .. ..$ range_022:List of 2
#>   .. .. ..$ negative_FOR: int [1:26] 313 314 315 316 317 318 319 320 321 322 ...
#>   .. .. ..$ positive_FOR: int [1:35] 666 667 668 669 670 671 672 673 674 675 ...
#>   .. ..$ range_023:List of 2
#>   .. .. ..$ negative_FOR: int [1:27] 312 313 314 315 316 317 318 319 320 321 ...
#>   .. .. ..$ positive_FOR: int [1:31] 664 665 666 667 668 669 670 671 672 673 ...
#>   .. ..$ range_024:List of 2
#>   .. .. ..$ negative_FOR: int [1:31] 312 313 314 315 316 317 318 319 320 321 ...
#>   .. .. ..$ positive_FOR: int [1:25] 667 668 669 670 671 672 673 674 675 676 ...
#>   .. ..$ range_025:List of 2
#>   .. .. ..$ negative_FOR: int [1:28] 314 315 316 317 318 319 320 321 322 323 ...
#>   .. .. ..$ positive_FOR: int [1:29] 668 669 670 671 672 673 674 675 676 677 ...
#>   .. ..$ range_026:List of 2
#>   .. .. ..$ negative_FOR: int [1:29] 313 314 315 316 317 318 319 320 321 322 ...
#>   .. .. ..$ positive_FOR: int [1:29] 669 670 671 672 673 674 675 676 677 678 ...
#>   .. ..$ range_027:List of 2
#>   .. .. ..$ negative_FOR: int [1:42] 309 310 311 312 313 314 315 316 317 318 ...
#>   .. .. ..$ positive_FOR: int [1:26] 668 669 670 671 672 673 674 675 676 677 ...
#>   .. ..$ range_028:List of 2
#>   .. .. ..$ negative_FOR: int [1:39] 306 307 308 309 310 311 312 313 314 315 ...
#>   .. .. ..$ positive_FOR: int [1:31] 664 665 666 667 668 669 670 671 672 673 ...
#>   .. ..$ range_029:List of 2
#>   .. .. ..$ negative_FOR: int [1:40] 309 310 311 312 313 314 315 316 317 318 ...
#>   .. .. ..$ positive_FOR: int [1:33] 663 664 665 666 667 668 669 670 671 672 ...
#>   .. ..$ range_030:List of 2
#>   .. .. ..$ negative_FOR: int [1:43] 306 307 308 309 310 311 312 313 314 315 ...
#>   .. .. ..$ positive_FOR: int [1:33] 663 664 665 666 667 668 669 670 671 672 ...
#>   .. ..$ range_031:List of 2
#>   .. .. ..$ negative_FOR: int [1:44] 306 307 308 309 310 311 312 313 314 315 ...
#>   .. .. ..$ positive_FOR: int [1:31] 663 664 665 666 667 668 669 670 671 672 ...
#>   .. ..$ range_032:List of 2
#>   .. .. ..$ negative_FOR: int [1:33] 305 306 307 308 309 310 311 312 313 314 ...
#>   .. .. ..$ positive_FOR: int [1:29] 664 665 666 667 668 669 670 671 672 673 ...
#>   .. ..$ range_033:List of 2
#>   .. .. ..$ negative_FOR: int [1:28] 311 312 313 314 315 316 317 318 319 320 ...
#>   .. .. ..$ positive_FOR: int [1:40] 665 666 667 668 669 670 671 672 673 674 ...
#>   .. ..$ range_034:List of 2
#>   .. .. ..$ negative_FOR: int [1:52] 306 307 308 309 310 311 312 313 314 315 ...
#>   .. .. ..$ positive_FOR: int [1:44] 663 664 665 666 667 668 669 670 671 672 ...
#>   .. ..$ range_035:List of 2
#>   .. .. ..$ negative_FOR: int [1:53] 302 303 304 305 306 307 308 309 310 311 ...
#>   .. .. ..$ positive_FOR: int [1:32] 661 662 663 664 665 666 667 668 669 670 ...
#>   .. ..$ range_036:List of 2
#>   .. .. ..$ negative_FOR: int [1:37] 300 301 302 303 304 305 306 307 308 309 ...
#>   .. .. ..$ positive_FOR: int [1:29] 666 667 668 669 670 671 672 673 674 675 ...
#>   .. ..$ range_037:List of 2
#>   .. .. ..$ negative_FOR: int [1:63] 301 302 303 304 305 306 307 308 309 310 ...
#>   .. .. ..$ positive_FOR: int [1:40] 664 665 666 667 668 669 670 671 672 673 ...
#>   .. ..$ range_038:List of 2
#>   .. .. ..$ negative_FOR: int [1:53] 305 306 307 308 309 310 311 312 313 314 ...
#>   .. .. ..$ positive_FOR: int [1:38] 659 660 661 662 663 664 665 666 667 668 ...
#>   .. ..$ range_039:List of 2
#>   .. .. ..$ negative_FOR: int [1:64] 297 298 299 300 301 302 303 304 305 306 ...
#>   .. .. ..$ positive_FOR: int [1:88] 610 611 612 613 614 615 616 617 618 619 ...
#>   .. ..$ range_040:List of 2
#>   .. .. ..$ negative_FOR: int [1:43] 300 301 302 303 304 305 306 307 308 309 ...
#>   .. .. ..$ positive_FOR: int [1:28] 669 670 671 672 673 674 675 676 677 678 ...
#>   .. ..$ range_041:List of 2
#>   .. .. ..$ negative_FOR: int [1:56] 298 299 300 301 302 303 304 305 306 307 ...
#>   .. .. ..$ positive_FOR: int [1:56] 610 611 612 613 614 615 616 617 618 619 ...
#>   .. ..$ range_042:List of 2
#>   .. .. ..$ negative_FOR: int [1:54] 298 299 300 301 302 303 304 305 306 307 ...
#>   .. .. ..$ positive_FOR: int [1:89] 610 611 612 613 614 615 616 617 618 619 ...
#>   .. ..$ range_043:List of 2
#>   .. .. ..$ negative_FOR: int [1:48] 301 302 303 304 305 306 307 308 309 310 ...
#>   .. .. ..$ positive_FOR: int [1:46] 610 611 612 613 614 615 616 617 618 619 ...
#>   .. ..$ range_044:List of 2
#>   .. .. ..$ negative_FOR: int [1:43] 302 303 304 305 306 307 308 309 310 311 ...
#>   .. .. ..$ positive_FOR: int [1:45] 610 611 612 613 614 615 616 617 618 619 ...
#>   .. ..$ range_045:List of 2
#>   .. .. ..$ negative_FOR: int [1:42] 303 304 305 306 307 308 309 310 311 312 ...
#>   .. .. ..$ positive_FOR: int [1:64] 610 611 612 613 614 615 616 617 618 619 ...
#>   .. ..$ range_046:List of 2
#>   .. .. ..$ negative_FOR: int [1:41] 303 304 305 306 307 308 309 310 311 312 ...
#>   .. .. ..$ positive_FOR: int [1:69] 625 626 627 628 629 630 631 632 633 634 ...
#>   .. ..$ range_047:List of 2
#>   .. .. ..$ negative_FOR: int [1:51] 301 302 303 304 305 306 307 308 309 310 ...
#>   .. .. ..$ positive_FOR: int [1:23] 658 659 660 661 662 663 664 665 666 667 ...
#>   .. ..$ range_048:List of 2
#>   .. .. ..$ negative_FOR: int [1:52] 301 302 303 304 305 306 307 308 309 310 ...
#>   .. .. ..$ positive_FOR: int [1:21] 661 662 663 664 665 666 667 668 669 670 ...
#>   .. ..$ range_049:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_050:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_051:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_052:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_053:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_054:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_055:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_056:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_057:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_058:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_059:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_060:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_061:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_062:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>   .. ..$ range_063:List of 2
#>   .. .. ..$ negative_FOR: int(0) 
#>   .. .. ..$ positive_FOR: int(0) 
#>  - attr(*, "MUSIC_data")= list()
#>  - attr(*, "NoiseLevel")=List of 3
#>   ..$ : num [1:63] 3.60e-11 3.71e-11 3.69e-11 3.84e-11 3.51e-11 ...
#>   ..$ : num [1:63] 2.39e-11 2.63e-11 2.67e-11 2.73e-11 2.54e-11 ...
#>   ..$ : num [1:63] 7.86e-11 8.65e-11 8.23e-11 8.25e-11 8.38e-11 ...
#>  - attr(*, "interpolated_doppler_cells_index")= int(0) 
#>  - attr(*, "reference_noise_normalized_limits_estimation_interval")=List of 2
#>   ..$ low_limit : num 0.9
#>   ..$ high_limit: num 1
#>  - attr(*, "class")= chr [1:2] "SeaSondeRCS" "list"
# }