Execute the MUSIC Algorithm on a SeaSondeRCS Object

This function performs the MUSIC (MUltiple SIgnal Classification) algorithm on a given SeaSondeRCS object, executing a series of processing steps to extract direction-of-arrival (DOA) information and other related metrics from the radar cross-spectrum data.

Usage

seasonder_runMUSIC(seasonder_cs_object)

Arguments

seasonder_cs_object

A SeaSondeRCS object that contains the radar cross-spectrum data and metadata. This object is modified in place to include the results of the MUSIC algorithm.

Value

A SeaSondeRCS object with updated MUSIC-related attributes. Specifically:

Processing steps annotated with the MUSIC start and end points.

Updated attributes and fields for covariance matrix computations, DOA estimations, and other MUSIC-related metrics.

Details

The MUSIC algorithm is executed in a series of sequential steps:

Log the start of the MUSIC algorithm.

Update the processing steps of the SeaSondeRCS object to include the MUSIC start text.

Perform the following computations:

Compute the covariance matrix from the cross-spectrum data.

Perform eigen decomposition on the covariance matrix.

Compute the DOA functions using MUSIC-specific methods.

Extract peaks from the DOA functions, corresponding to possible signal directions.

Calculate the signal power matrix.

Test for dual solutions and compute their proportions.

Select the final set of DOAs from the computed data.

Convert the selected DOAs to geographical coordinates (latitude and longitude).

Log the completion of the MUSIC algorithm.

See also

Examples

# \donttest{
# Prepare a SeaSondeRCS object with MUSIC data
apm_file <- system.file("css_data/MeasPattern.txt", package = "SeaSondeR")
apm_obj <- seasonder_readSeaSondeRAPMFile(apm_file)
#> seasonder_createSeaSondeRAPM:  APM object created successfully.
cs_file <- system.file("css_data/CSS_TORA_24_04_04_0700.cs", package = "SeaSondeR")
cs_obj <- seasonder_createSeaSondeRCS(cs_file, seasonder_apm_object = apm_obj)
#> new_SeaSondeRCS:  SeaSondeRCS object created successfully.
cs_obj <- SeaSondeR:::seasonder_initMUSICData(cs_obj)
cs_obj <- SeaSondeR:::seasonder_initMUSICData(
 cs_obj,
 range_cells = c(rep(5,11), rep(4,11)),
 doppler_bins = c(c(669:679),c(674:684))
)
cs_obj <- SeaSondeR:::seasonder_SeaSondeRCSMUSICInterpolateDoppler(cs_obj)
# Run the MUSIC algorithm
cs_obj <- seasonder_runMUSIC(cs_obj)
#> seasonder_runMUSIC:  MUSIC algorithm started.
#> seasonder_runMUSIC:  MUSIC algorithm finished.
# Check the updated processing steps
print(seasonder_getSeaSondeRCS_ProcessingSteps(cs_obj))
#>  [1] "2025-04-24 22:23:22.412135: FOR parameters set."                                                                       
#>  [2] "2025-04-24 22:23:22.423984: NoiseLevel for antenna 1 computed."                                                        
#>  [3] "2025-04-24 22:23:22.438239: NoiseLevel for antenna 2 computed."                                                        
#>  [4] "2025-04-24 22:23:22.452552: NoiseLevel for antenna 3 computed."                                                        
#>  [5] "2025-04-24 22:23:22.460271: Created from /home/runner/work/_temp/Library/SeaSondeR/css_data/CSS_TORA_24_04_04_0700.cs."
#>  [6] "2025-04-24 22:23:22.524523: MUSIC algorithm started."                                                                  
#>  [7] "2025-04-24 22:23:22.524778: MUSIC covariance matrix computation started"                                               
#>  [8] "2025-04-24 22:23:22.52798: MUSIC covariance matrix computation ended"                                                  
#>  [9] "2025-04-24 22:23:22.530323: NoiseLevel for antenna 1 computed."                                                        
#> [10] "2025-04-24 22:23:22.532532: NoiseLevel for antenna 2 computed."                                                        
#> [11] "2025-04-24 22:23:22.53466: NoiseLevel for antenna 3 computed."                                                         
#> [12] "2025-04-24 22:23:22.538464: MUSIC covariance matrix decomposition started"                                             
#> [13] "2025-04-24 22:23:22.540111: MUSIC covariance matrix decomposition ended"                                               
#> [14] "2025-04-24 22:23:22.540324: MUSIC DOA functions computation started"                                                   
#> [15] "2025-04-24 22:23:22.567927: MUSIC DOA functions computation ended"                                                     
#> [16] "2025-04-24 22:23:22.568149: MUSIC peak extraction started"                                                             
#> [17] "2025-04-24 22:23:22.582971: MUSIC peak extraction ended"                                                               
#> [18] "2025-04-24 22:23:22.585887: MUSIC dual solutions testing started"                                                      
#> [19] "2025-04-24 22:23:22.600529: MUSIC dual solutions testing ended"                                                        
#> [20] "2025-04-24 22:23:22.600771: MUSIC DOA selection started"                                                               
#> [21] "2025-04-24 22:23:22.601935: MUSIC DOA selection ended"                                                                 
#> [22] "2025-04-24 22:23:22.606119: MUSIC algorithm ended with 42.9% of dual solutions."                                       
# }