Validate Attributes for a SeaSondeRAPM Object

This function validates the attributes of a given SeaSondeRAPM object to ensure they meet the required specifications.

Usage

seasonder_validateAttributesSeaSondeRAPM(seasonde_apm_obj)

Arguments

seasonde_apm_obj

A SeaSondeRAPM object whose attributes are to be validated.

Value

TRUE if all attributes are valid. The function will stop execution and display an error message if any of the attributes are invalid.

Details

The function performs validation on the following attributes of the SeaSondeRAPM object:

quality_matrix

BEAR

Type

Creator

SiteName

SiteOrigin

FileName

CreateTimeStamp

ProcessingSteps

AmplitudeFactors

AntennaBearing

StationCode

BearingResolution

Smoothing

CommentLine

FileID

PhaseCorrections

It internally calls specific validation functions for each of these attributes. If any of the attributes are found to be invalid, the function will stop execution and display an error message.

For more details on the attributes, see seasonder_initializeAttributesSeaSondeRAPM.

See also

Examples

# Create a mock SeaSondeRAPM object with valid attributes
mock_matrix <- matrix(complex(real = c(1, 2), imaginary = c(3, 4)), nrow = 2)
mock_apm <- seasonder_createSeaSondeRAPM(mock_matrix)
#> seasonder_createSeaSondeRAPM:  APM object created successfully.

# Validate attributes
valid <- seasonder_validateAttributesSeaSondeRAPM(mock_apm)
print(valid)  # Should print TRUE if all attributes are valid
#> [1] TRUE