Validate First Order Region (FOR) Parameters

This function validates and assigns default values to the parameters used in defining the First Order Region (FOR) in a SeaSondeR cross-spectral object. It ensures that all necessary parameters are present and assigns appropriate defaults where values are missing.

Usage

seasonder_validateFOR_parameters(
  seasonder_cs_obj,
  FOR_parameters,
  method = "SeaSonde"
)

Arguments

seasonder_cs_obj

A SeaSondeRCS object containing metadata about the Doppler spectrum.

FOR_parameters

A named list containing the parameters for first-order region detection.

method

A character string specifying the validation method. Default is "SeaSonde". Currently, only "SeaSonde" is supported.

Value

A named list containing validated and completed FOR parameters.

Details

The function validates FOR parameters and assigns default values where necessary. If the selected method is "SeaSonde", the function ensures that each parameter is defined and, if missing, assigns it a default value based on seasonder_defaultFOR_parameters.

The parameters validated include:

nsm (Doppler Smoothing): Number of points used for spectral smoothing.

fdown (Peak Power Dropoff): Defines how far below peak power the algorithm descends before searching for the null.

flim (Null Below Peak Power): Specifies a power threshold for identifying the first-order region.

noisefact (Signal to Noise): Threshold above the noise floor that a spectral bin must exceed to be considered first-order.

currmax (Maximum Velocity): Maximum radial velocity allowed in the first-order region.

reject_distant_bragg: Logical flag indicating whether to reject Bragg regions that are too distant from the central Bragg frequency.

reject_noise_ionospheric: Logical flag indicating whether to reject Bragg regions affected by ionospheric noise.

reject_noise_ionospheric_threshold: Threshold (in dB) for rejecting first-order regions based on noise contamination.

reference_noise_normalized_limits: Estimated reference noise range in normalized Doppler frequency, computed using seasonder_estimateReferenceNoiseNormalizedLimits.

See also

Examples

if (FALSE) { # \dontrun{
# Validate and complete FOR parameters for a SeaSondeRCS object
validated_params <- seasonder_validateFOR_parameters(cs_obj, list(fdown = 12))
print(validated_params)
} # }