Eigen Decomposition of the MUSIC Covariance Matrix

Performs the eigen decomposition of a MUSIC covariance matrix to obtain the eigenvalues and eigenvectors. This decomposition is a critical step in the MUSIC algorithm for spectral estimation and direction finding.

Usage

seasonder_MUSICCovDecomposition(seasonder_cs_object)

Arguments

C

A 3x3 complex covariance matrix obtained from the seasonder_MUSICComputeCov function.

Value

A list containing the eigenvalues and eigenvectors of the covariance matrix. The eigenvalues are sorted in descending order to facilitate their use in MUSIC algorithm applications. The list has the following components:

values: A numeric vector containing the sorted eigenvalues.

vectors: A 3x3 matrix of the corresponding eigenvectors in columns, reordered to match the sorted eigenvalues.

Details

The function uses the eigen function to perform the eigen decomposition of the covariance matrix. It then sorts the eigenvalues in descending order, as this ordering is often required for further analysis in MUSIC algorithm applications, such as identifying the signal and noise subspaces. The eigenvectors are reordered accordingly to maintain the correspondence with their eigenvalues.