Calculate Euclidean Distances for MUSIC Algorithm

Computes the Euclidean distances based on the eigenvalue decomposition from the MUSIC covariance matrix and the antenna pattern measurements (APM) for single and dual solutions.

Usage

seasonder_MUSICEuclideanDistance(seasonder_cs_object)

Arguments

eigen_analysis

A list containing the results of the eigenvalue decomposition of the MUSIC covariance matrix. This list must have at least two elements: $values for eigenvalues and $vectors for eigenvectors.

seasonder_apm_obj

A matrix representing the antenna pattern measurements (APM), where each column corresponds to a specific bearing and contains the complex-valued response of the antenna at that bearing.

Value

A 2xN complex matrix, where N is the number of bearings. The first row ("single") contains the Euclidean distances for the single solution scenario, and the second row ("dual") for the dual solution scenario, based on the eigenvalue decomposition and the antenna pattern measurements.

Details

The Euclidean distance in the context of the MUSIC algorithm is a measure used to estimate the direction of arrival (DOA) of signals. This function computes these distances by projecting the antenna pattern measurements onto the subspace orthogonal to the signal's eigenvectors (from the eigenvalue decomposition of the MUSIC covariance matrix). The computation for single and dual solutions involves the use of eigenvectors corresponding to the noise subspace (En). For each bearing, the function calculates the projected Euclidean distance using the formula:

$$D_i = a^H (En E_n^H) a$$

where \(a\) represents the complex-valued response of the antenna at a given bearing, \(En\) the matrix of eigenvectors corresponding to the noise subspace, and \(H\) denotes the conjugate transpose. The distances are calculated for both single and dual solutions, indicating scenarios with different numbers of signal sources.