Get Doppler Bins Frequency

This function calculates the frequency limits for each Doppler bin within a SeaSonde Cross Spectrum (CS) object. It can return frequencies either in their original Hz values or normalized by the second Bragg frequency. The frequencies are calculated as the high limit of each Doppler bin interval, similar to what is displayed in SpectraPlotterMap.

Usage

seasonder_getDopplerBinsFrequency(seasonder_cs_obj, normalized = FALSE)

Arguments

seasonder_cs_obj

A SeaSonde Cross Spectrum (CS) object created by seasonder_createSeaSondeRCS(). This object contains the necessary metadata and spectral data to compute Doppler bin frequencies.

normalized

A logical value indicating if the returned frequencies should be normalized by the second Bragg frequency. When TRUE, frequencies are divided by the second Bragg frequency, returning dimensionless values relative to it. Default is FALSE, returning frequencies in Hz.

Value

A numeric vector of frequencies representing the high limit of each Doppler bin interval. If normalized is TRUE, these frequencies are dimensionless values relative to the second Bragg frequency; otherwise, they are in Hz.

Details

The function internally utilizes several helper functions such as seasonder_getCenterDopplerBin(), seasonder_getnDopplerCells(), and seasonder_getDopplerSpectrumResolution() to calculate the Doppler bin frequencies. Furthermore, when normalization is requested, it uses seasonder_getBraggDopplerAngularFrequency() to obtain the second Bragg frequency for normalization purposes.