Initialize Cross-Spectra Data Structure for SeaSondeR

This function initializes a data structure for storing cross-spectra data related to SeaSonde radar measurements. It creates a list of matrices, each corresponding to different components of the SeaSonde data.

Usage

seasonder_initCSDataStructure(nRanges, nDoppler)

Arguments

nRanges

Integer, number of range cells in the radar measurement. Specifies the number of rows in each matrix.

nDoppler

Integer, number of Doppler bins in the radar measurement. Specifies the number of columns in each matrix.

Value

A list containing matrices for different cross-spectra components:

SSA1: Matrix for SSA1 component, filled with NA_real_.

SSA2: Matrix for SSA2 component, filled with NA_real_.

SSA3: Matrix for SSA3 component, filled with NA_real_.

CS12: Matrix for CS12 component, complex numbers with NA_real_ real and imaginary parts.

CS13: Matrix for CS13 component, complex numbers with NA_real_ real and imaginary parts.

CS23: Matrix for CS23 component, complex numbers with NA_real_ real and imaginary parts.

QC: Quality control matrix, filled with NA_real_.