Initialize Attributes for a SeaSondeRAPM Object

This function initializes attributes for a SeaSondeRAPM object, including metadata and properties.

Usage

seasonder_initializeAttributesSeaSondeRAPM(calibration_matrix, ...)

Arguments

calibration_matrix

A 2 x b complex matrix, where b is the number of bearings for calibration.

...

Additional named attributes that may override the defaults.

Value

A list containing initialized attributes for a SeaSondeRAPM object.

Details

The function initializes the following attributes:

quality_matrix: A 2 x b complex matrix for quality data, where b is the number of bearings.

BEAR: A numeric vector for bearings (degrees).

Type: Character string for antenna pattern type.

Creator: Object creator name. Default is an empty character vector.

SiteName: Site name (not the same as SiteCode). Default is an empty character vector.

SiteOrigin: Numeric vector with two elements representing the Station GPS location. Default is c(0,0).

FileName: Default is an empty character vector.

CreateTimeStamp: APM file creation time. Default is current system date and time.

ProcessingSteps: Processing steps applied to this object. Default is an empty character vector.

AmplitudeFactors: Numeric vector with two elements for the amplitude factors. Default is c(0,0).

AntennaBearing: Default is an empty numeric vector.

StationCode: 4-character station code. Default is an empty character vector.

BearingResolution: In degrees. Default is an empty numeric vector.

Smoothing: Numeric vector indicating smoothing applied to the antenna pattern. Default is an empty numeric vector.

CommentLine: Metadata lines in the data file not matching any other attribute. Default is an empty character vector.

FileID: File's UUID. Default is an empty character vector.

PhaseCorrections: Numeric vector with two elements for phase corrections. Default is c(0,0).

Each attribute has a corresponding setter and getter function, following the naming pattern seasonder_set/getSeaSondeRAPM_*, where * is the name of the attribute. For example, to set the 'Type' attribute, you would use seasonder_setSeaSondeRAPM_Type(seasonde_apm_obj,new_value). To get the same, you would use seasonder_getSeaSondeRAPM_Type(seasonde_apm_obj,new_value).

Please see seasonder_validateAttributesSeaSondeRAPM for details in attributes validation.

See also

Examples

# Create a calibration_matrix
cal_matrix <- matrix(complex(real = c(1, 2), imaginary = c(3, 4)), nrow = 2, ncol = 2)

# Initialize attributes with default settings
attr_list <- seasonder_initializeAttributesSeaSondeRAPM(calibration_matrix = cal_matrix)
str(attr_list)
#> List of 17
#>  $ quality_matrix   : cplx [1:2, 1:2] NA NA NA ...
#>  $ BEAR             : num [1:2] 0 0
#>  $ Type             : chr(0) 
#>  $ Creator          : chr(0) 
#>  $ SiteName         : chr(0) 
#>  $ SiteOrigin       : num [1:2] 0 0
#>  $ FileName         : chr(0) 
#>  $ CreateTimeStamp  : POSIXct[1:1], format: "2024-05-27 08:17:54"
#>  $ ProcessingSteps  : chr(0) 
#>  $ AmplitudeFactors : num [1:2] 0 0
#>  $ AntennaBearing   : num(0) 
#>  $ StationCode      : chr(0) 
#>  $ BearingResolution: num(0) 
#>  $ Smoothing        : num(0) 
#>  $ CommentLine      : chr(0) 
#>  $ FileID           : chr(0) 
#>  $ PhaseCorrections : num [1:2] 0 0

# Initialize attributes with custom 'Type'
attr_list_custom <- seasonder_initializeAttributesSeaSondeRAPM(calibration_matrix = cal_matrix,
Type = "Custom Pattern")
str(attr_list_custom)
#> List of 17
#>  $ quality_matrix   : cplx [1:2, 1:2] NA NA NA ...
#>  $ BEAR             : num [1:2] 0 0
#>  $ Type             : chr "Custom Pattern"
#>  $ Creator          : chr(0) 
#>  $ SiteName         : chr(0) 
#>  $ SiteOrigin       : num [1:2] 0 0
#>  $ FileName         : chr(0) 
#>  $ CreateTimeStamp  : POSIXct[1:1], format: "2024-05-27 08:17:54"
#>  $ ProcessingSteps  : chr(0) 
#>  $ AmplitudeFactors : num [1:2] 0 0
#>  $ AntennaBearing   : num(0) 
#>  $ StationCode      : chr(0) 
#>  $ BearingResolution: num(0) 
#>  $ Smoothing        : num(0) 
#>  $ CommentLine      : chr(0) 
#>  $ FileID           : chr(0) 
#>  $ PhaseCorrections : num [1:2] 0 0