Read Specifications from a YAML File

This function reads a YAML file containing specifications, handles potential reading errors, and extracts specific information based on a provided path.

Usage

seasonder_readYAMLSpecs(file_path, path)

Arguments

file_path

A string. The path to the YAML file.

path

A character vector. Represents the path within the YAML file to access the desired information. For example, to access fields of version V2 of the header, the path would be c("header", "versions", "V2").

Value

A list. The information extracted from the YAML file based on the provided path.

Error Handling

The function has built-in error handling which aborts the function's execution and logs detailed error messages in case of:

File not found.

Error in reading the YAML content.

If the read YAML content is not of list type.

If no data is found for the provided path in the YAML content.

Errors generated are of class "seasonder_read_yaml_file_error". Detailed error information including the file path and path within the file is provided. For logging and aborting, this function utilizes the seasonder_logAndAbort function.

See also

Examples

if (FALSE) { # \dontrun{
# Assuming a YAML file named "example.yaml" exists with appropriate content
result <- seasonder_readYAMLSpecs("example.yaml", c("header", "versions", "V2"))
print(result)
} # }